+++ Riot/Modules/Authentication/AuthenticationViewController.m
@@ -18,7 +18,6 @@
 
 #import "AuthenticationViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 #import "MXSession+Riot.h"