+++ Riot/Modules/BugReport/BugReportViewController.m
@@ -16,7 +16,6 @@
 
 #import "BugReportViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 
 #import "GBDeviceInfo_iOS.h"