+++ Riot/Modules/Communities/Members/GroupParticipantsViewController.m
@@ -16,7 +16,6 @@
 
 #import "GroupParticipantsViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 
 #import "Contact.h"