+++ Riot/Modules/Contacts/Details/ContactDetailsViewController.m
@@ -18,7 +18,6 @@
 
 #import "ContactDetailsViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 #import "MXSession+Riot.h"