+++ Riot/Modules/GlobalSearch/Messages/HomeMessagesSearchViewController.m
@@ -17,7 +17,6 @@
 
 #import "HomeMessagesSearchViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 
 #import "HomeViewController.h"