+++ Riot/Modules/Integrations/Widgets/WidgetViewController.m
@@ -17,7 +17,6 @@
 
 #import "WidgetViewController.h"
 
-#import "AppDelegate.h"
 #import "IntegrationManagerViewController.h"
 #import "Riot-Swift.h"