+++ Riot/Modules/Room/Members/Detail/RoomMemberDetailsViewController.m
@@ -18,7 +18,6 @@
 
 #import "RoomMemberDetailsViewController.h"
 
-#import "AppDelegate.h"
 #import "Riot-Swift.h"
 
 #import "RoomMemberTitleView.h"