+++ Riot/Modules/Room/RoomViewController.m
@@ -21,8 +21,6 @@
 #import "RoomDataSource.h"
 #import "RoomBubbleCellData.h"
 
-#import "AppDelegate.h"
-
 #import "RoomInputToolbarView.h"
 #import "DisabledRoomInputToolbarView.h"
 
@@ -4525,7 +4523,7 @@
         {
             // The room has an active jitsi widget
             // Show it in the banner if the user is not already in
-            AppDelegate *appDelegate = [AppDelegate theDelegate];
+            LegacyAppDelegate *appDelegate = [AppDelegate theDelegate];
             if ([appDelegate.jitsiViewController.widget.widgetId isEqualToString:jitsiWidget.widgetId])
             {
                 if ([self checkUnsentMessages] == NO)