+++ Riot/Modules/Settings/Security/SecurityViewController.m
@@ -22,7 +22,6 @@
 
 #import <OLMKit/OLMKit.h>
 
-#import "AppDelegate.h"
 #import "AvatarGenerator.h"
 
 #import "ThemeService.h"