+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changes to be released in next version
  * 
     
 🙌 Improvements
- * 
+ * AppDelegate: Convert to Swift (#3594).
     
 🐛 Bugfix
  *